requirements.txt
# FastAPI and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# OpenAI
openai==1.3.9

# Database
pymongo==4.6.1
motor==3.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
python-json-logger==2.0.7
